requirements.txt
mariadb>=1.1.10
python-dotenv>=1.0.1
httpx>=0.27.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
passlib>=1.7.4
cryptography>=42.0.0
